move TcpMessage to Mmcc.Bot.Polychat.Networking
