+++ src/Mmcc.Bot.Polychat/MessageHandlers/HandleCommandResult.cs
@@ -7,6 +7,7 @@ using System.Threading.Tasks;
 using MediatR;
 using Microsoft.Extensions.Logging;
 using Mmcc.Bot.Common.Models.Colours;
+using Mmcc.Bot.Polychat.Networking;
 using Remora.Discord.API.Abstractions.Rest;
 using Remora.Discord.API.Objects;
 using Remora.Rest.Core;
