+++ src/Mmcc.Bot.Polychat/MessageHandlers/HandlePlayersOnlineMessage.cs
@@ -2,6 +2,7 @@
 using System.Linq;
 using MediatR;
 using Mmcc.Bot.Polychat.Abstractions;
+using Mmcc.Bot.Polychat.Networking;
 using Mmcc.Bot.Polychat.Services;
 
 namespace Mmcc.Bot.Polychat.MessageHandlers;
