+++ src/Mmcc.Bot.Polychat/MessageHandlers/HandleServerInfo.cs
@@ -2,6 +2,7 @@
 using Microsoft.Extensions.Logging;
 using Mmcc.Bot.Polychat.Abstractions;
 using Mmcc.Bot.Polychat.Models;
+using Mmcc.Bot.Polychat.Networking;
 using Mmcc.Bot.Polychat.Services;
 
 namespace Mmcc.Bot.Polychat.MessageHandlers;
