+++ src/Mmcc.Bot.Polychat/MessageHandlers/HandleServerStatus.cs
@@ -8,6 +8,7 @@ using Microsoft.Extensions.Logging;
 using Mmcc.Bot.Common.Models.Colours;
 using Mmcc.Bot.Polychat.Abstractions;
 using Mmcc.Bot.Polychat.Models.Settings;
+using Mmcc.Bot.Polychat.Networking;
 using Mmcc.Bot.Polychat.Services;
 using Remora.Discord.API.Abstractions.Rest;
 using Remora.Discord.API.Objects;
